+++ .scalafmt.conf
@@ -1,2 +1,5 @@
 version = 3.5.8
 runner.dialect = scala213
+
+# Don't keep the params lined up with each other.
+align.openParenDefnSite = false
