+++ src/main/scala/org/renci/babel/validator/Reporter.scala
@@ -1,0 +1,143 @@
+package org.renci.babel.validator
+
+import com.typesafe.scalalogging.LazyLogging
+import org.renci.babel.validator.Validator.Conf
+import org.renci.babel.validator.model.{BabelOutput, Compendium}
+import zio.ZIO
+import zio.blocking.Blocking
+import zio.console.Console
+import zio.stream.ZStream
+
+import java.io.{FileOutputStream, PrintStream}
+
+/**
+ * Functions for reporting on the differences between two input files.
+ */
+object Reporter extends LazyLogging {
+  /**
+   * Helper method for displaying the percent change between two counts.
+   */
+  def relativePercentChange(count: Long, countPrev: Long): String = {
+    val percentChange = (count - countPrev).toDouble / countPrev * 100
+    f"${count - countPrev}%+d\t$percentChange%+2.2f%%"
+  }
+
+  /**
+   * Generic method to determine whether a particular filename should be
+   * filtered in or out from the results. The algorithm we use is:
+   * 1. If any `--filtered-in` prefixes are provided, then we exclude everything
+   *    that isn't explicitly filtered in (by starting with one of those prefixes
+   *    in a case-sensitive manner).
+   * 2. Otherwise, all filenames are allowed EXCEPT those explicitly filtered out
+   *    by `--filtered-out` by starting with one of those prefixes in a
+   *    case-sensitive manner.
+   */
+  def filterFilename(conf: Conf, filename: String): Boolean = {
+    val filteredIn = conf.filterIn.getOrElse(List())
+    val filteredOut = conf.filterOut.getOrElse(List())
+
+    if (filteredIn.nonEmpty) {
+      if (filteredIn.exists(filename.startsWith(_))) {
+        return true;
+      } else {
+        return false;
+      }
+    }
+
+    if (filteredOut.nonEmpty && filteredOut.exists(filename.startsWith(_))) {
+      return false;
+    }
+
+    true
+  }
+
+  /**
+   * Given two BabelOutputs, it returns a list of all compendia found in BOTH of the BabelOutputs
+   * paired together.
+   *
+   * TODO: modify this so we return every compendium found in EITHER BabelOutput.
+   */
+  def retrievePairedCompendiaSummaries(
+    babelOutput: BabelOutput,
+    babelPrevOutput: BabelOutput
+  ): Seq[(String, Compendium, Compendium)] = {
+    for {
+      summary <- babelOutput.compendia
+      summaryPrev <- babelPrevOutput.compendia
+      if summaryPrev.filename == summary.filename
+    } yield {
+      (summary.filename, summary, summaryPrev)
+    }
+  }
+
+  def diffResults(conf: Conf): ZIO[Blocking with Console, Throwable, Unit] = {
+    val babelOutput = new BabelOutput(conf.babelOutput())
+    val babelPrevOutput = new BabelOutput(conf.babelPrevOutput())
+    val output = conf.output.toOption match {
+      case Some(file) => new PrintStream(new FileOutputStream(file))
+      case _          => System.out
+    }
+
+    val pairedSummaries =
+      retrievePairedCompendiaSummaries(babelOutput, babelPrevOutput)
+    output.println("Filename\tCount\tPrevCount\tDiff\tPercentageChange")
+    ZStream
+      .fromIterable(pairedSummaries)
+      .mapMParUnordered(conf.nCores()) {
+        case (
+          filename: String,
+          summary: Compendium,
+          prevSummary: Compendium
+          ) if filterFilename(conf, filename) => {
+          for {
+            count <- summary.count
+            prevCount <- prevSummary.count
+            typesChunk <- (for {
+              row: Compendium.Record <- summary.records
+            } yield (row.`type`)).runCollect
+            typesErrors <- summary.types.catchAll(err => {
+              logger.error(s"Types error: ${err}")
+              ZIO.fail(err)
+            })
+            prevTypesChunk <- (for {
+              row: Compendium.Record <- prevSummary.records
+            } yield (row.`type`)).runCollect
+
+            // types <- summary.typesZIO
+            // prevTypes <- prevSummary.typesZIO
+          } yield {
+            output.println(
+              s"${filename}\t${count}\t${prevCount}\t${relativePercentChange(count, prevCount)}"
+            )
+
+            val types = typesChunk.toSet
+            val prevTypes = prevTypesChunk.toSet
+
+            val added = types -- prevTypes
+            val deleted = prevTypes -- types
+            val changeString = (added.toSeq, deleted.toSeq) match {
+              case (Seq(), Seq()) => "No change"
+              case (added, Seq()) => s"Added: ${added}"
+              case (Seq(), _) => s"Deleted: ${added}"
+              case (added, deleted) =>
+                s"Added: ${added}, Deleted: ${deleted}"
+            }
+
+            output.println(
+              s"${filename}\t${types.mkString(", ")} (${typesChunk.length})\t${
+                prevTypes
+                  .mkString(", ")
+              } (${prevTypesChunk.length})\t${changeString}"
+            )
+          }
+        }
+        case (filename: String, _, _) if !filterFilename(conf, filename) => {
+          logger.info(s"Skipping ${filename}")
+          ZIO.succeed(())
+        }
+        case abc =>
+          ZIO.fail(new RuntimeException(s"Invalid paired summary: ${abc}"))
+      }
+      .runDrain
+  }
+}
