+++ src/main/scala/org/renci/babel/validator/model/Compendium.scala
@@ -1,6 +1,7 @@
 package org.renci.babel.validator.model
 
 import com.typesafe.scalalogging.LazyLogging
+import org.renci.babel.validator.model.Compendium.{Identifier, Record}
 import zio.ZIO
 import zio.blocking.Blocking
 import zio.stream._
@@ -9,26 +10,50 @@ import zio.json._
 import java.io.File
 import scala.collection.mutable
 
-// Q&D memorize from https://stackoverflow.com/a/36960228/27310
-object Compendium {
+object Compendium extends LazyLogging {
+  /**
+   * Quick-and-dirty memoize() implementation from https://stackoverflow.com/a/36960228/27310
+   * This should probably be replaced with ZIO Cache or ScalaCache at some point.
+   *
+   * @param f The function to memoize.
+   * @tparam I The input type
+   * @tparam O The output type
+   * @return A function that will either return the cached value or calculate and cache it.
+   */
   def memoize[I, O](f: I => O): I => O = new mutable.HashMap[I, O]() {
-    override def apply(key: I) = getOrElseUpdate(key, f(key))
+    override def apply(key: I) = {
+      logger.debug(s"Caching ${f}(${key}), already cached: ${contains(key)}")
+      getOrElseUpdate(key, f(key))
+    }
   }
+
+  /** An identifier in this compendium. */
+  case class Identifier(
+    i: Option[String],
+    l: Option[String]
+  )
+
+  /** A single record in this compendium. */
+  case class Record(
+    `type`: String,
+    ic: Option[Double],
+    identifiers: Seq[Identifier]
+  )
 }
 
-/** A Compendium models a single compendium in a Babel output.
-  *
-  * At the moment, this is a JSON object with the following structure: { "type":
-  * "biolink:...", "identifiers": [{ "i": "identifier", "l": "label" }, { ... }]
-  * }
-  *
-  * Since these files can be VERY large, we should only process them in a stream
-  * if we need to generate any kind of summary statistics or to verify things.
-  */
+/**
+ * A Compendium models a single compendium in a Babel output.
+ *
+ * At the moment, this is a JSON object with the following structure:
+ * { "type": "biolink:...", "identifiers": [{ "i": "identifier", "l": "label" }, { ... }] }
+ */
 class Compendium(file: File) extends LazyLogging {
   val filename = file.getName
   val path = file.toPath
 
+  /**
+   * A ZStream of all the lines in this file as strings.
+   */
   lazy val lines: ZStream[Blocking, Throwable, String] = {
     ZStream
       .fromFile(path)
@@ -36,32 +61,29 @@ class Compendium(file: File) extends LazyLogging {
       .aggregate(ZTransducer.splitLines)
   }
 
-  case class Identifier(
-      i: Option[String],
-      l: Option[String]
-  )
-
-  case class CompendiumRecord(
-      `type`: String,
-      ic: Option[Double],
-      identifiers: Seq[Identifier]
-  )
-
+  /* Implicit decoders for parts of the Record. */
   implicit val identifierDecoder: JsonDecoder[Identifier] =
     DeriveJsonDecoder.gen[Identifier]
-  implicit val recordDecoder: JsonDecoder[CompendiumRecord] =
-    DeriveJsonDecoder.gen[CompendiumRecord]
+  implicit val recordDecoder: JsonDecoder[Record] =
+    DeriveJsonDecoder.gen[Record]
 
-  lazy val recordsRaw
-      : ZStream[Blocking, Throwable, Either[String, CompendiumRecord]] = {
-    lines.map(line => line.fromJson[CompendiumRecord])
+  /**
+   * A ZStream that _doesn't_ throw an exception when you go through the entries: instead,
+   * any record that can't be converted to a Record is kept as an error as a String.
+   */
+  lazy val recordsRaw: ZStream[Blocking, Throwable, Either[String, Record]] = {
+    lines.map(line => line.fromJson[Record])
   }
 
-  lazy val records: ZStream[Blocking, Throwable, CompendiumRecord] = {
+  /**
+   * A ZStream of all the compendium records in this file. Throws an exception if any
+   * line could not be converted into a String.
+   */
+  lazy val records: ZStream[Blocking, Throwable, Record] = {
     lines
       .flatMap(line =>
         line
-          .fromJson[CompendiumRecord]
+          .fromJson[Record]
           .fold(
             err =>
               ZStream.fail(
@@ -72,28 +94,10 @@ class Compendium(file: File) extends LazyLogging {
       )
   }
 
-  // TODO: get rid of Summary, replace with direct calls to the wrapped object
-  case class Summary(
-      filename: String,
-      file: File,
-      countZIO: ZIO[Blocking, Throwable, Long],
-      typesZIO: ZIO[Blocking, Throwable, Set[String]],
-      typesZStream: ZStream[
-        Blocking,
-        Throwable,
-        Either[String, CompendiumRecord]
-      ]
-  )
-
-  def summary: Summary = Summary(
-    filename,
-    file,
-    count,
-    types,
-    recordsRaw
-  )
-
+  /** Count the total number of lines in this file. */
   def count: ZIO[Blocking, Throwable, Long] = lines.runCount
+
+  /** Returns the Set of all the unique types in this file. */
   def types: ZIO[Blocking, Throwable, Set[String]] =
     records.map(_.`type`).fold(Set[String]())(_ + _)
 }
